cleanup & redis not required
