+++ Back/ecoreleve_server/__init__.py
@@ -14,12 +14,12 @@ from .Models import (
     BaseExport,
     dbConfig,
     db,
-    loadThesaurusTrad,
     groupfinder
 )
 from .Views import add_routes, add_cors_headers_response_callback
 from pyramid.events import NewRequest
 from sqlalchemy.orm import sessionmaker, scoped_session
+from .utils.loadThesaurus import loadThesaurusTrad
 
 
 def datetime_adapter(obj, request):
@@ -86,7 +86,8 @@ def main(global_config, **settings):
     config.include('pyramid_tm')
     config.include('pyramid_jwtauth')
 
-    config.registry.dbmaker = scoped_session(sessionmaker(bind=engine, autoflush=False))
+    config.registry.dbmaker = scoped_session(
+        sessionmaker(bind=engine, autoflush=False))
     dbConfig['dbSession'] = scoped_session(sessionmaker(bind=engine))
     config.add_request_method(db, name='dbsession', reify=True)
